refactor: turn project to typescript

+++ contracts/interfaces/Weth9.sol
@@ -1,15 +1,14 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-
 interface WETH9 {
     /** ETH swap to WETH */
     function deposit() external payable;
 
     /** WETH swap to ETH */
-    function withdraw(uint wad) external;
+    function withdraw(uint256 wad) external;
 
-    function approve(address guy, uint wad) external returns(bool);
+    function approve(address guy, uint256 wad) external returns (bool);
 
-    function balanceOf(address owner) external view returns(uint);
+    function balanceOf(address owner) external view returns (uint256);
 }
